devをmainにデプロイ

+++ server/main.py
@@ -7,6 +7,7 @@ from os.path import join, dirname
 from flask_jwt_extended import JWTManager, create_access_token
 from pymongo import MongoClient
 from Auth import Auth
+from crypto.generate import generate_ed25519_keypair
 
 
 # Settings インスタンス
@@ -29,6 +30,9 @@ jwt = JWTManager(app)
 client = MongoClient("localhost", 27017)
 db = client["db"]
 users = db["users"]
+device_keys = db["device_keys"]
+pre_pairings = db["pre_pairings"]
+pairings = db["pairings"]
 
 
 @app.route('/signup', methods=['POST'])
@@ -125,6 +129,15 @@ def delete_account():
     if not user:
         return jsonify({'error': 'Invalid token'}), 400
 
+    email = user['email']
+    # パスワードの認証
+    firebase_api_key = settings.firebase_api_key
+    res = Auth.sign_in_with_email_and_password(
+        firebase_api_key, email, password)
+
+    if 'error' in res:
+        return jsonify({'error': res['error']['message']}), 400
+
     try:
         auth.delete_user(user['uid'])
     except Exception as e:
@@ -133,6 +146,106 @@ def delete_account():
     users.delete_many({'uid': user['uid']})
 
     return jsonify({'done': 'success'}), 200
+
+
+@app.route('/generate_device_key', methods=['POST'])
+def generate_device_key():
+    uid = request.json['uid']
+    if not uid:
+        return jsonify({'error': 'Missing uid'}), 400
+
+    if device_keys.find_one({'uid': uid}):
+        return jsonify({'error': 'Device key already exists'}), 400
+
+    private_key, public_key = generate_ed25519_keypair()
+
+    device_keys.insert_one({
+        'uid': uid,
+        'private_key': private_key,
+        'public_key': public_key
+    })
+
+    return jsonify({'private_key': private_key, 'public_key': public_key}), 200
+
+
+@app.route('/register_pairing', methods=['POST'])
+def register_pairing():
+    token = request.json['token']
+    if not token:
+        return jsonify({'error': 'Missing token'}), 400
+    public_key = request.json['public_key']
+    if not public_key:
+        return jsonify({'error': 'Missing public_key'}), 400
+
+    user = users.find_one({'token': token})
+    if not user:
+        return jsonify({'error': 'Invalid token'}), 400
+
+    device_key = device_keys.find_one({'public_key': public_key})
+    if not device_key:
+        return jsonify({'error': 'Invalid public_key'}), 400
+
+    pre_pairings.insert_one({
+        'uid': user['uid'],
+        'public_key': public_key
+    })
+
+    return jsonify({'done': 'pre_pairing'}), 200
+
+
+@app.route('/check_pairing', methods=['POST'])
+def check_pairing():
+    token = request.json['token']
+    if not token:
+        return jsonify({'error': 'Missing token'}), 400
+    private_key = request.json['private_key']
+    if not private_key:
+        return jsonify({'error': 'Missing private_key'}), 400
+
+    user = users.find_one({'token': token})
+    if not user:
+        return jsonify({'error': 'Invalid token'}), 400
+
+    device_key = device_keys.find_one({'private_key': private_key})
+    if not device_key:
+        return jsonify({'error': 'Invalid private_key'}), 400
+
+    pre_pairing = pre_pairings.find_one(
+        {'uid': user['uid'], 'public_key': device_key['public_key']})
+
+    if not pre_pairing:
+        return jsonify({'error': 'Invalid pairing'}), 400
+
+    pre_pairings.delete_many(
+        {'uid': user['uid'], 'public_key': device_key['public_key']})
+
+    pairings.delete_many({'uid': user['uid']})
+    pairings.delete_many({'public_key': device_key['public_key']})
+    pairings.delete_many({'private_key': private_key})
+
+    pairings.insert_one({
+        'uid': user['uid'],
+        'public_key': device_key['public_key'],
+        'private_key': private_key
+    })
+
+    return jsonify({'done': 'success'}), 200
+
+
+@app.route('/auth_check', methods=['POST'])
+def auth_check():
+    token = request.json['token']
+    if not token:
+        return jsonify({'error': 'Missing token'}), 400
+
+    user = users.find_one({'token': token})
+    if not user:
+        return jsonify({'error': 'Invalid token'}), 400
+
+    new_token = create_access_token(identity=user['token'])
+    users.update_one({'token': token}, {'$set': {'token': new_token}})
+
+    return jsonify({'token': new_token}), 200
 
 
 if __name__ == '__main__':
